A couple small changed for the Pomodoro timer, including separating the frames into separate files so recording is easier.
